pyproject.toml
[tool.poetry]
name = "smorest-sfs"
version = "0.1.0"
description = ""
authors = ["ssfdust <ssfdust@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
flask-smorest = "^0.21.0"
sqlalchemy-utils = { git = "https://github.com/yk-lab/sqlalchemy-utils.git" }
sqlalchemy = "^1.3.12"
flask-sqlalchemy = "^2.4.1"
marshmallow-sqlalchemy = "^0.22.3"
flask-login = "^0.4.1"
kombu = "^4.6.7"
toml = "^0.10.0"
flask-mail = "^0.9.1"
flask-marshmallow = "^0.11.0"
loguru = "^0.4.1"
pyperclip = "^1.7.0"
tabulate = "^0.8.6"
sqlparse = "^0.3.1"
Flask-Babel = "^1.0.0"
passlib = "^1.7.2"
flask-jwt-extended = "^3.24.1"
captcha = "^0.3"
flask-redis = "^0.4.0"
alembic = "^1.4.2"
flask-migrate = "^2.5.3"
pendulum = "^2.1.0"
rope = "^0.16.0"
sqlalchemy-mptt = "^0.2.5"
openpyxl = "^3.0.3"
gunicorn = "^20.0.4"
meinheld = "^1.0.1"
celery = "^4.4.2"
sqlalchemy-mixins = "^1.2.1"
psycopg2-binary = "^2.8.5"
celery-redbeat = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = { git = "https://github.com/pytest-dev/pytest.git" }
pdbpp = "^0.10.2"
freezegun = "^0.3.13"
pytest-cov = "^2.8.1"
ipython = "^7.13.0"
pylint = "^2.4.4"
mypy = "^0.770"
invoke = "^1.4.0"
sqlalchemy-stubs = "^0.3"
sqlalchemy-mptt-stubs = "^0.0.3"
supervisor = "^4.1.0"
deepdiff = "^4.3.2"

[tool.coverage.run]
include = ["smorest_sfs"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
